Implement New AnimateAsState composable function
