+++ app/src/commonMain/kotlin/com/henni/handwriting/ui/extensions/ColorPickerExtension.kt
@@ -1,4 +1,4 @@
-package com.henni.handwriting.ui
+package com.henni.handwriting.ui.extensions
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -14,15 +14,12 @@ import androidx.compose.material.icons.filled.Check
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.henni.handwriting.ui.extensions.VerticalSpacer
 
 @Composable
 fun HandWritingColorPicker(
