+++ app/src/commonMain/kotlin/com/henni/handwriting/ui/extensions/SliderExtension.kt
@@ -23,7 +23,6 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.henni.handwriting.ui.SliderRange
 
 @Composable
 fun HandWritingSwitch(
